Change customTooltip to tooltip, remove sistentheme layer from code, added triggers, html support in guidance.js
