--- src/pages/projects/sistent/components/customTooltip/code.js
@@ -1,8 +1,0 @@
-import React from "react";
-import CustomTooltipCode from "../../../../../sections/Projects/Sistent/components/customTooltip/code";
-
-const CustomTooltipCodePage = () => {
-  return <CustomTooltipCode />;
-};
-
-export default CustomTooltipCodePage;
