--- src/pages/projects/sistent/components/customTooltip/guidance.js
@@ -1,8 +1,0 @@
-import React from "react";
-import CustomTooltipGuidance  from "../../../../../sections/Projects/Sistent/components/customTooltip/guidance";
-
-const CustomTooltipGuidancePage = () => {
-  return <CustomTooltipGuidance />;
-};
-
-export default CustomTooltipGuidancePage;
