--- src/pages/projects/sistent/components/customTooltip/index.js
@@ -1,8 +1,0 @@
-import React from "react";
-import SistentCustomTooltip from "../../../../../sections/Projects/Sistent/components/customTooltip";
-
-const SistentCustomTooltipPage = () => {
-  return <SistentCustomTooltip />;
-};
-
-export default SistentCustomTooltipPage;
