+++ src/pages/projects/sistent/components/tooltip/code.js
@@ -1,0 +1,8 @@
+import React from "react";
+import TooltipCode from "../../../../../sections/Projects/Sistent/components/tooltip/code";
+
+const TooltipCodePage = () => {
+  return <TooltipCode />;
+};
+
+export default TooltipCodePage;
