+++ src/pages/projects/sistent/components/tooltip/guidance.js
@@ -1,0 +1,8 @@
+import React from "react";
+import TooltipGuidance  from "../../../../../sections/Projects/Sistent/components/tooltip/guidance";
+
+const TooltipGuidancePage = () => {
+  return <TooltipGuidance />;
+};
+
+export default TooltipGuidancePage;
