+++ src/pages/projects/sistent/components/tooltip/index.js
@@ -1,0 +1,8 @@
+import React from "react";
+import SistentTooltip from "../../../../../sections/Projects/Sistent/components/tooltip";
+
+const SistentTooltipPage = () => {
+  return <SistentTooltip />;
+};
+
+export default SistentTooltipPage;
